# Reduce lock-contention in ServeMux