@tmthrgd 2018-11-26
@miekg Do you know where any particular hot spots are in the compression code?