@miekg 2018-11-27
What are we optimizing here for exactly. If mutex scale badly on large machines I can see that being an issue, but this adds a whole bunch of code for what end result exactly? 

Compression uses an expansive map and probably a whole lot of smaller things. I saw your PRs I'll ttake a look