@tmthrgd 2018-11-27
I think I'd favour simplicity here for now. This approach is a counter to the poor scaling of `sync.RWMutex` on many cored systems, while also offering a performance improvement in the hot-path. I'd say it's fine to close both this and #809 until there are more specific complaints about the performance of `ServeMux` (keeping in mind it can be re-implemented externally with either of these PRs applied).

@miekg Ack, I've addressed some of the compression performance pitfalls and there should now be a very nice improvement. It may be possible to do better from there. (I'll somewhat hijack #709 for that).