@miekg 2019-01-09
don't think this or #809 will be merged.

Shall I close both?